refactor: reemplazar layout global por shell nuevo (topbar + rail)
- Layout viejo (dark sidebar) reemplazado completamente
- Nuevo shell: topbar global (logo + search + sync + user) + rail colapsado
- Aplica a toda la app: dashboard, procesar, importar, facturas, entrenamiento
- Búsqueda global llama a /api/invoices/search
- Tokens verde musgo/gris aplicados globalmente
- Mobile-first: drawer off-canvas expandible en topbar
